fix: Fixed some minor spelling errors.
There are a few things to note about this commit:

First, the corrections use American english spellings.  This affects
things like 'cancelled' (British) vs. 'canceled' (American).  I don't
know which dictionary is being used by the project to catch spelling
errors, so you may wish to revert those changes.

Second, there are words in use like 'deschedule' that aren't actually in
the English language or any of its dialects.  However, there is no
better word than that, so even if your dictionary flags it in the
future, I'd suggest that you keep using it (the word makes sense to
native English speakers).

Finally, there are a few grammar errors that I've avoided changing as
they are not spelling errors, and therefore outside of the scope of this
commit.